# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ManageHR API",
        "description": "Documentación de la API RESTful de ManageHR desarrollada en Laravel con autenticación JWT",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/tipoContrato": {
            "get": {
                "tags": [
                    "Tipos de Contrato"
                ],
                "summary": "Listar todos los tipos de contrato",
                "operationId": "20f7f75599bd4ee83bf3028223928744",
                "responses": {
                    "200": {
                        "description": "Lista completa de tipos de contrato",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tipocontrato": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tipos de Contrato"
                ],
                "summary": "Intento de creación de tipo de contrato (no permitido)",
                "operationId": "4b96b911f2c23531edda48cb47eb48b3",
                "responses": {
                    "400": {
                        "description": "Este módulo no permite crear registros"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tipoContrato/{id}": {
            "get": {
                "tags": [
                    "Tipos de Contrato"
                ],
                "summary": "Obtener un tipo de contrato por ID",
                "operationId": "371408f070cd68a02df98b77dc5ce522",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tipo de contrato",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo de contrato encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tipocontrato": {
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Tipo de contrato no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tipos de Contrato"
                ],
                "summary": "Intento de actualización de tipo de contrato (no permitido)",
                "operationId": "c3ebf30c3723a623c2813b48b1737732",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tipo de contrato",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Este módulo no permite actualizar registros"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tipos de Contrato"
                ],
                "summary": "Intento de eliminación de tipo de contrato (no permitido)",
                "operationId": "5db77533be7547c965fce49a9c247f9e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tipo de contrato",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Este módulo no permite eliminar registros"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Tipos de Contrato"
                ],
                "summary": "Intento de actualización parcial de tipo de contrato (no permitido)",
                "operationId": "05b2539db315840bbffd9dda44277a01",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tipo de contrato",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Este módulo no permite actualizar registros"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Introduce tu token JWT en el formato: Bearer {token}"
            }
        }
    },
...
}